utils/job_status/requirements.txt
google-cloud-logging_v2==3.5.0
google-cloud-storage